Re-organized items.js and created seperate routes for account settings and fridge models
