+++ back-end/routes/items.js
@@ -194,120 +194,6 @@ router.post('/scan', (req, res) => {
   });
 });
 
-/**
- * @route   GET /Account-Setting/:field
- * @desc    Get the field of the account setting
- * @access  Public
- */
-router.get("/Account-Setting/:field", (req, res) => {
-  try {
-    const { field } = req.params;
-    
-    // Define allowed fields
-    const allowedFields = ["name", "email", "phone"];
-    
-    if (!allowedFields.includes(field)) {
-      return res.status(400).json({ error: "Invalid field" });
-    }
-    
-    //for database
-    // const user = await User.findById(req.user.id);
-    
-    // Mock data for example
-    const userSettings = {
-      name: "John Doe",
-      email: "john@example.com",
-      phone: "1234567890",
-    };
-    
-    // Check if the field exists for this user
-    if (userSettings[field] === undefined) {
-      return res.status(404).json({ error: `${field} not found` });
-    }
-    
-    // Return just the requested field
-    res.json({ [field]: userSettings[field] });
-    
-  } catch (error) {
-    console.error(`Error fetching ${req.params.field}:`, error);
-    res.status(500).json({ error: "Server error while fetching user data" });
-  }
-});
-
-/**
- * @route   GET /Fridge-Model
- * @desc    Get all fridge model settings
- * @access  Public
- */
-router.get("/Fridge-Model", (req, res) => {
-  console.log("🔵 Fridge-Model route was hit!");
-  try {
-    // Mock data for example
-    const fridgeSettings = {
-      brand: "Samsung",
-      model: "S29",
-      features: {
-        humidity: true,
-        freezer: true,
-        vegetableDrawer: false
-      }
-    };
-    
-    res.json(fridgeSettings);
-    
-  } catch (error) {
-    console.error("Error fetching fridge settings:", error);
-    res.status(500).json({ error: "Server error while fetching fridge data" });
-  }
-});
-
-/**
- * @route   POST /Fridge-Model
- * @desc    Update fridge model settings
- * @access  Public
- */
-router.post("/Fridge-Model", (req, res) => {
-  try {
-    const { brand, model, features } = req.body;
-    
-    // Validate required fields
-    if (!brand || !model) {
-      return res.status(400).json({ error: "Brand and model are required" });
-    }
-    
-    // Here you would typically save to database
-    // For now, just return the received data as confirmation
-    res.json({
-      brand,
-      model,
-      features: features || {}
-    });
-    
-  } catch (error) {
-    console.error("Error updating fridge settings:", error);
-    res.status(500).json({ error: "Server error while updating fridge data" });
-  }
-});
-
-
-/**
- * @route   POST /Account-Setting/:field
- * @desc    Get the field of the account setting
- * @access  Public
- */
-router.post("/Account-Setting/:field", (req,res)=> {
-  try{
-    const { field } = req.params;
-    const { value } = req.body;
-
-    res.json({ [field]: value });
-  }
-  catch (error) {
-    console.error("Error updating user field:", error);
-    res.status(500).json({ error: "Server error while updating user data" });
-  }
-})
-
 // add frequent items endpoint
 router.get('/frequent', (req, res) => {
   const frequentItems = [
